+++ CHANGELOG.md
@@ -21,6 +21,7 @@ The format of this changelog is based on [Keep a Changelog](https://keepachangel
 
 * Properly hide unavailable load conditions on Classic [#223]
 * Fix macro too long message sometimes incorrectly showing upon login
+* Improve startup behavior to ensure bindings are always loaded correctly even in combat
 
 ## [1.16.2] - 2024-09-05